db_config.py
import mysql.connector
def get_db_connection():
    return mysql.connector.connect(
        host="localhost",
        user="root",
        password="Tarun_143",
        database="student_db1"
    )
