requirements.txt
FLASK==3.0.3
mysql-connector-python=9.0.0